backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
python-dotenv==1.0.1
PyMySQL==1.1.0
SQLAlchemy==2.0.23
bcrypt==4.1.2
